% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_section_files.R
\name{bind_section_files}
\alias{bind_section_files}
\title{Bind each section's files together}
\usage{
bind_section_files(folder_name, export_csv = FALSE)
}
\arguments{
\item{folder_name}{A character vector containing the name of the folders the
results are stored in. CSV files stored in the same folder are bound together
and exported into a CSV file that has the same name as the path of the folder
they were stored in.}

\item{export_csv}{A logical vector, defaults to \code{TRUE}. If \code{TRUE},
the concatenated tibbles are saved in a "file" folder in the folder indicated
by \code{folder_name}}
}
\value{
A list containing the section files.
}
\description{
Binds the files for each section into one tibble
}
\examples{
bind_section_files("test_scrape")

}
