% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_main_suffixes.R
\name{get_full_section_subs}
\alias{get_full_section_subs}
\title{Get entire main section}
\usage{
get_full_section_subs(main_section_suffix, folder_name = NULL)
}
\arguments{
\item{main_section_suffix}{The section's suffix}

\item{folder_name}{A character vector containing the folder name the scraped
files are supposed to be stored in}
}
\value{
A tibble with two columns: the sub(sub)section's suffix and a folder
name based on the given folder name and the sub(sub)section's name
}
\description{
Returns a file that allows for scraping the entire section, folder names
mimic the section structure
}
\examples{
get_full_section_subs(main_section_suffix = "/f102", folder_name = NULL)

}
