% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_thread_links.R
\name{get_full_thread_links}
\alias{get_full_thread_links}
\title{Get entire main section's links}
\usage{
get_full_thread_links(
  suffix,
  path,
  cut_off = "2000-01-01",
  delay = TRUE,
  export_links = FALSE,
  export_meta = TRUE,
  output_folder = ""
)
}
\arguments{
\item{suffix}{The section's suffix}

\item{cut_off}{A character string containing the date at which the latest
post in the thread should had been posted on. Has to be in the format
\code{YYYY-MM-DD}. Defaults to \code{"2000-01-01"}.}

\item{delay}{flashback.org's robots.txt-file asks for putting a five
second delay between each iteration. You can deliberately ignore this by
setting \code{delay = FALSE}. Note that THIS IS NOT RECOMMENDED!}

\item{export_links}{If set to \code{TRUE}, a CSV file containing the links is
exported}

\item{export_meta}{If set to \code{TRUE}, a CSV file containing data on the
scrape is exported}

\item{output_folder}{A character string determining the folder the CSV files
containing the links and the meta data should be stored in.}

\item{folder_name}{A character vector with a folder name the scraped files
are supposed to be stored in}
}
\value{
A tibble with the name of the sub(sub) section's suffix
\code{sub_suffix}, the name of the folder the scraped thread should be stored
at \code{folder_name}, the thread links \code{thread_links}, and the
prospective file name \code{file_name}
}
\description{
Returns a file that allows for scraping the entire section, folder names
mimic the section structure
}
\examples{
get_full_section_links(suffix = "/f102", folder_name = NULL,
  cut_off = "2020-10-25", delay = TRUE)

}
