% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sections.R
\name{get_main_sections}
\alias{get_main_sections}
\title{Get main sections}
\usage{
get_main_sections(mainpage = "https://www.flashback.org")
}
\arguments{
\item{mainpage}{the flashback.org homepage; default value suffices}
}
\value{
a tibble with two columns: 'name' and 'suffix'
}
\description{
Returns the main sections' names and suffixes
}
\examples{
get_main_sections()

}
