% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_threads.R
\name{get_missing_threads}
\alias{get_missing_threads}
\title{Get missing urls}
\usage{
get_missing_threads(link_vec, folder_name)
}
\arguments{
\item{link_vec}{A vector containing the links that should have been scraped}

\item{folder_name}{A character vector with a folder name the scraped files
are stored in}
}
\value{
The link tibble containing the links that haven't been scraped so far
}
\description{
If the process has stopped, you can read in the files and determine the
missing ones
}
\examples{
get_full_section_subs(link_tbl = links_science, folder_name = "science")

}
