% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_threads.R
\name{get_missing_urls}
\alias{get_missing_urls}
\title{Get missing urls}
\usage{
get_missing_urls(link_tbl, folder_name)
}
\arguments{
\item{link_tbl}{A tibble containing the thread links that are to be scraped;
thread links need to be stored in a column named \code{thread_links}}

\item{folder_name}{A character vector with a folder name the scraped files
are stored in}
}
\value{
The link tibble containing the links that haven't been scraped so far
}
\description{
If the process has stopped, you can read in the files and determine the
missing ones
}
\examples{
get_full_section_subs(link_tbl = links_science, folder_name = "science")

}
