% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_sections.R
\name{get_sub}
\alias{get_sub}
\title{Get sub-sections}
\usage{
get_sub(main_section_suffix)
}
\arguments{
\item{main_section_suffix}{a string containing a main section's suffix
(which can be obtained using \code{\link{get_main_sections}})}
}
\value{
a character vector containing the sub-sections suffixes.
Please note that sub-sections can also contain sub-sub-sections.
These can be obtained using \code{\link{get_subsub}}
}
\description{
Returns the sub-sections' names and suffixes
}
\examples{
get_sub("/f4")

}
