% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_sub_sections.R
\name{get_subsub}
\alias{get_subsub}
\title{Get sub-sub-sections}
\usage{
get_subsub(sub_section_suffix)
}
\arguments{
\item{sub_section_suffix}{a string containing a sub-section's suffix
(which can be obtained using \code{\link{get_sub_sections}}). Suffixes need
to start with '/'.}
}
\value{
a tibble with two columns: `sub_section` contains the sub-section's
suffix and `sub_sub_section` the sub-sub-sections suffixes.
If there are no sub-sub-sections, `sub_sub_section` is `NULL`.
}
\description{
Returns the sub-sub-sections' names suffixes
}
\examples{
get_subsub("/f243")

}
