% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thread_links.R
\name{get_thread_links}
\alias{get_thread_links}
\title{Get thread links in section}
\usage{
get_thread_links(
  suffix,
  cut_off = "2000-01-01",
  title = TRUE,
  delay = TRUE,
  pure_suffix = FALSE
)
}
\arguments{
\item{suffix}{A character string containing a sub-section's suffix
(which can be obtained using \code{\link{get_sub_sections}}). Suffixes need
to start with \code{/}.}

\item{cut_off}{A character string containing the date at which the latest
post in the thread should had been posted on. Has to be in the format
\code{YYYY-MM-DD}. Defaults to \code{"2000-01-01"}.}

\item{title}{A logical vector. If \code{TRUE}, the thread title is returned
as well.}

\item{delay}{flashback.org's robots.txt-file asks for putting a five
second delay between each iteration. You can deliberately ignore this by
setting \code{delay = FALSE}. Note that THIS IS NOT RECOMMENDED!}
}
\value{
A character vector containing the links. If \code{title = TRUE}, a
tibble which contains the thread's title and URL.
}
\description{
Returns the links of threads within a section
}
\examples{
get_thread_links(suffix = "/f245", cut_off = "2020-05-01", title = TRUE, delay = TRUE)

}
