% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thread_titles.R
\name{get_thread_titles}
\alias{get_thread_titles}
\title{Get thread titles in section}
\usage{
get_thread_titles(suffix, delay = TRUE)
}
\arguments{
\item{suffix}{A character string containing a sub-section's suffix
(which can be obtained using \code{\link{get_sub_sections}}). Suffixes need
to start with \code{/}.}

\item{delay}{flashback.org's robots.txt-file asks for putting a five
second delay between each iteration. You can deliberately ignore this by
setting \code{delay = FALSE}. Note that THIS IS NOT RECOMMENDED!}
}
\value{
A tibble containing the links, titles, and dates.
}
\description{
Returns the links of threads within a section
}
\examples{
get_thread_links(suffix = "/f245", cut_off = "2020-05-01", title = TRUE, delay = TRUE)

}
