% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_thread_content.R
\name{scrape_thread_content}
\alias{scrape_thread_content}
\title{Scrape thread}
\usage{
scrape_thread_content(
  suffix,
  export_csv = FALSE,
  folder_name = NULL,
  file_name = NULL,
  delay = TRUE
)
}
\arguments{
\item{suffix}{A character string containing a thread's suffix (which can be
obtained using \code{\link{get_thread_links()}}). Suffixes need to start with
\code{/}.}

\item{export_csv}{A logical vector. Defaults to \code{FALSE}. The function
can automatically save the output in a csv file. If \code{export_csv = TRUE}
, a csv file is exported. The output folder can be specified using the
\code{folder} argument.}

\item{folder_name}{A character string which specifies the name of the folder
the output should be saved in. The folder's name is added to the path of the
current working directory which can be obtained using \code{getwd()} and
modified with \code{setwd()}. If nothing is specified and
\code{export_csv = TRUE}, the function will export the csv file straight into
the working directory.}

\item{file_name}{A character string which specifies the name of the output
file. It is not necessary to add `.csv`. If no file name is provided,
\code{file_name} defaults to \code{scrape_[YYYY-MM-DD].csv}.}

\item{delay}{A logical vector, defaults to \code{TRUE}. flashback.org's
robots.txt-file asks for putting a five second delay between each iteration.
You can deliberately ignore this by setting \code{delay = FALSE}. Note that
THIS IS NOT RECOMMENDED!}
}
\value{
A tibble with the following columns: \code{url} contains the thread's
URL suffix, \code{date} the date the posting was made on, \code{time} the
time the posting was made at, \code{author_name} the respective author's user
name, \code{author_url} the link to their profile (can be scraped using
\code{scrape_user_profile()}), \code{quoted_user} the user name of the user
that is quoted in a posting (\code{NA} if the posting does not contain a
quote), \code{posting} the posting *as is*, i.e., with potential quotes,
\code{posting_wo_quote} the posting with all quotes removed.
}
\description{
Scrapes a certain thread
}
\examples{
scrape_thread_content(suffix = "/t3145103", export_csv = TRUE, folder_name = "sandbox/results", file_name = "test", delay = FALSE)

}
