% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_user_profiles.R
\name{scrape_user_profile}
\alias{scrape_user_profile}
\title{Scrape profiles}
\usage{
scrape_user_profile(suffix)
}
\arguments{
\item{suffix}{A character string containing a user profile's suffix.}
}
\value{
A tibble containing the user name, the date a user joined, their status, and the number of postings.
}
\description{
Scrapes a certain user profile
}
\examples{
scrape_user_profile(suffix = "/u1000001")

}
