% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_threads.R
\name{update_threads}
\alias{update_threads}
\title{Update section file}
\usage{
update_threads(last_scrape, section_suffix, old_tibble)
}
\arguments{
\item{last_scrape}{A character vector indicating the date the urls were
scraped for the last time}

\item{section_suffix}{The suffix for the main section the newer threads are
to be taken from}

\item{old_tbl}{The former scrape}
}
\value{
The old tibble, updated with the new postings
}
\description{
New thread posts need to be updated once in a while
}
\examples{
update_threads(last_scrape = "2020-11-11", section_suffix = "/f8", old_tibble = old_tbl)

}
