% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sections.R
\name{get_main_sections}
\alias{get_main_sections}
\title{Get main sections}
\usage{
get_main_sections(mainpage = "https://www.flashback.org")
}
\arguments{
\item{mainpage}{The flashback.org homepage; default value suffices.}
}
\value{
A tibble with two columns: 'name' and 'suffix'. The former indicates
the section's name, the latter its link's suffix.
}
\description{
Returns the main sections' names and suffixes
}
\examples{
get_main_sections()

}
