% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_sections.R
\name{get_sub}
\alias{get_sub}
\title{Get sub-sections}
\usage{
get_sub(main_section_suffix)
}
\arguments{
\item{main_section_suffix}{A character string containing a main section's
suffix (which can be obtained using \code{\link{get_main_sections}}). Needs
to start with \code{/}.}
}
\value{
A tibble containing the sub-sections' titles and suffixes. Please
note that sub-sections can also contain sub-sub-sections. These can be
obtained using \code{\link{get_subsub}}.
}
\description{
Returns the sub-sections' names and link suffixes
}
\examples{
get_sub("/f4")

}
