% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_sub_sections.R
\name{get_subsub}
\alias{get_subsub}
\title{Get sub-sub-sections}
\usage{
get_subsub(sub_section_suffix)
}
\arguments{
\item{sub_section_suffix}{A string containing a sub-section's suffix
(which can be obtained using \code{\link{get_sub_sections}}). Suffixes need
to start with '/'.}
}
\value{
A tibble with the sub-sub-sections' titles and suffixes. If there are
no sub-sub-sections, it is `NULL` and a warning is printed.
}
\description{
Returns the sub-sub-sections' names and link suffixes
}
\examples{
get_subsub("/f243")

}
